+++ example/hello-static/webapp/main.dart
@@ -4,5 +4,5 @@ library hello_static;
 import "package:stream/stream.dart";
 
 void main() {
-  new StreamServer().run();
+  new StreamServer().start();
 }