+++ test/syntax/syntax.rsp.dart
@@ -7,8 +7,7 @@ import "package:stream/stream.dart";
 
 /** Template, syntax, for rendering the view. */
 void syntax(HttpConnect connect, {foo, bool c:false}) { //9
-  final request = connect.request, response = connect.response,
-    output = response.outputStream;
+  final request = connect.request, response = connect.response;
   var _v_;
   response.headers.contentType = new ContentType.fromString("""${foo.contentType}""");
 
@@ -16,34 +15,34 @@ void syntax(HttpConnect connect, {foo, bool c:false}) { //9
 
   response.headers.add("accept-ranges", foo.acceptRanges); //#9
 
-  output.writeString("""
+  response.addString("""
 <!DOCTYPE html>
 <html>
   <head>
     <title>"""); //#10
 
   _v_ = "$foo.name [${foo.title}]"; //#13
-  if (_v_ != null) output.writeString("$_v_");
+  if (_v_ != null) response.addString("$_v_");
 
   //#13
-  output.writeString("""
+  response.addString("""
 </title>
   </head>
   <body>
     <p>This is a test with """);
-  output.writeString('"""');
-  output.writeString("""
+  response.addString('"""');
+  response.addString("""
 .
     <p>Another expresion: """);
 
   _v_ = foo.description; //#17
-  if (_v_ != null) output.writeString("$_v_");
+  if (_v_ != null) response.addString("$_v_");
 
-  output.writeString("""
+  response.addString("""
 
     <p>An empty expression: """); //#17
 
-  output.writeString("""
+  response.addString("""
 
     <p>This is not a tag: [foo ], [another and [/none].
     <ul>
@@ -51,40 +50,40 @@ void syntax(HttpConnect connect, {foo, bool c:false}) { //9
 
   for (var user in foo.friends) { //#22
 
-    output.writeString("""      <li>"""); //#23
+    response.addString("""      <li>"""); //#23
 
     _v_ = user.name; //#23
-    if (_v_ != null) output.writeString("$_v_");
+    if (_v_ != null) response.addString("$_v_");
 
-    output.writeString("""
+    response.addString("""
 
 """); //#23
 
     if (user.isCustomer) { //#24
 
-      output.writeString("""
+      response.addString("""
       <i>!important!</i>
 """); //#25
     } //if
 
     while (user.hasMore()) { //#27
 
-      output.writeString("""        """); //#28
+      response.addString("""        """); //#28
 
       _v_ = user.showMore(); //#28
-      if (_v_ != null) output.writeString("$_v_");
+      if (_v_ != null) response.addString("$_v_");
 
-      output.writeString("""
+      response.addString("""
 
 """); //#28
     } //while
 
-    output.writeString("""
+    response.addString("""
       </li>
 """); //#30
   } //for
 
-  output.writeString("""
+  response.addString("""
     </ul>
 
 """); //#32
@@ -92,13 +91,13 @@ void syntax(HttpConnect connect, {foo, bool c:false}) { //9
   for (var fruit in ["apple", "orange"]) { //#34
   } //for
 
-  output.writeString("""
+  response.addString("""
 
 """); //#36
 
   if (foo.isCustomer) { //#37
 
-    output.writeString("""
+    response.addString("""
       *Custmer*
 """); //#38
 
@@ -109,7 +108,7 @@ void syntax(HttpConnect connect, {foo, bool c:false}) { //9
 
   } else if (foo.isEmployee) { //#41
 
-    output.writeString("""
+    response.addString("""
       *Employee*
 """); //#42
 
@@ -118,12 +117,12 @@ void syntax(HttpConnect connect, {foo, bool c:false}) { //9
 
   } else { //#44
 
-    output.writeString("""
+    response.addString("""
       *Unknown* [/if] 
 """); //#45
   } //if
 
-  output.writeString("""
+  response.addString("""
   </body>
 </html>